+++ src/main/java/com/example/apiformatech/service/SessionService.java
@@ -2,6 +2,7 @@ package com.example.apiformatech.service;
 
 import com.example.apiformatech.model.Establishment;
 import com.example.apiformatech.model.Session;
+import com.example.apiformatech.exception.ResourceNotFoundException;
 import com.example.apiformatech.repository.EstablishmentRepository;
 import com.example.apiformatech.repository.SessionRepository;
 import org.springframework.stereotype.Service;
@@ -23,6 +24,9 @@ public class SessionService {
 
     // Méthode pour sauvegarder une session
     public Session saveSession(Session session) {
+        if (sessionRepository.existsByName(session.getName())) {
+            throw new ResourceNotFoundException("Une session avec le même nom existe déjà.");
+        }
         return sessionRepository.save(session);
     }
 
@@ -38,15 +42,15 @@ public class SessionService {
 
     // Méthode pour associer une session à un établissement
     public Session assignSessionToEstablishment(Long sessionId, Long establishmentId) {
-        Session session = sessionRepository.findById(sessionId).orElseThrow(() -> new RuntimeException("Session not found"));
-        Establishment establishment = establishmentRepository.findById(establishmentId).orElseThrow(() -> new RuntimeException("Establishment not found"));
+        Session session = sessionRepository.findById(sessionId).orElseThrow(() -> new ResourceNotFoundException("Session not found"));
+        Establishment establishment = establishmentRepository.findById(establishmentId).orElseThrow(() -> new ResourceNotFoundException("Establishment not found"));
         session.setEstablishment(establishment);
         return sessionRepository.save(session);
     }
     // Méthode pour mettre à jour une session de formation
     public Session updateSession(Long id, Session updatedSession) {
         Session existingSession = sessionRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Session not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Session not found"));
 
         // Met à jour les informations de la session
         existingSession.setType(updatedSession.getType());
@@ -60,6 +64,9 @@ public class SessionService {
 
     // Méthode pour supprimer une session
     public void deleteSession(Long id) {
+        if (!sessionRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Session avec l'ID" + id + " n'existe pas");
+        }
         sessionRepository.deleteById(id);
     }
